+++ CHANGELOG.md
@@ -21,6 +21,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - matio 1.5.21 (VS 2022 x86, x64 build),
   - all others windows dependencies rebuilt with VS 2022.
 
+- Nelson uses `JuliaMono-Regular` font as default.
+
 - [#505](http://github.com/Nelson-numerical-software/nelson/issues/505): libCurl 7.81 on Windows.
 - [#524](http://github.com/Nelson-numerical-software/nelson/issues/524): oneApi 2022.1 on Windows.